refactor(app): 12 refactored to JS
