+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/root/Personal/http-please/src/composite.js":"1","/root/Personal/http-please/src/helpers.js":"2","/root/Personal/http-please/src/helpers.test.js":"3","/root/Personal/http-please/src/http-please.js":"4","/root/Personal/http-please/src/index.js":"5","/root/Personal/http-please/src/index.test.js":"6","/root/Personal/http-please/src/methods.js":"7","/root/Personal/http-please/vite.config.js":"8"},{"size":2084,"mtime":1692038915855,"results":"9","hashOfConfig":"10"},{"size":640,"mtime":1691796278199,"results":"11","hashOfConfig":"10"},{"size":1226,"mtime":1691796294559,"results":"12","hashOfConfig":"10"},{"size":1553,"mtime":1691794740669,"results":"13","hashOfConfig":"10"},{"size":59,"mtime":1691792268949,"results":"14","hashOfConfig":"10"},{"size":2031,"mtime":1691792339939,"results":"15","hashOfConfig":"10"},{"size":4715,"mtime":1692039221034,"results":"16","hashOfConfig":"10"},{"size":539,"mtime":1691793728369,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","suppressedMessages":"20","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"eawqg4",{"filePath":"21","messages":"22","suppressedMessages":"23","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","suppressedMessages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","suppressedMessages":"29","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","suppressedMessages":"32","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","suppressedMessages":"35","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","suppressedMessages":"38","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","suppressedMessages":"41","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/root/Personal/http-please/src/composite.js",[],[],"/root/Personal/http-please/src/helpers.js",[],[],"/root/Personal/http-please/src/helpers.test.js",[],[],"/root/Personal/http-please/src/http-please.js",[],[],"/root/Personal/http-please/src/index.js",[],[],"/root/Personal/http-please/src/index.test.js",[],[],"/root/Personal/http-please/src/methods.js",[],[],"/root/Personal/http-please/vite.config.js",[],["42"],{"ruleId":"43","severity":2,"message":"44","line":15,"column":22,"nodeType":"45","messageId":"46","endLine":15,"endColumn":31,"suppressions":"47"},"no-undef","'__dirname' is not defined.","Identifier","undef",["48"],{"kind":"49","justification":"50"},"directive",""]
