--- src/index.js
@@ -1,1 +1,0 @@
-export { CreateHttpPlease as default } from './composite';
