+++ tsconfig.json
@@ -2,13 +2,14 @@
   // Change this to match your project
   "include": ["src/**/*"],
   "exclude": [
-    "node_modules",
-    "src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}"
+    "node_modules"
+    // "src/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts,jsx,tsx}"
   ],
   "compilerOptions": {
     // Tells TypeScript to read JS files, as
     // normally they are ignored as source files
     "allowJs": true,
+    "checkJs": true,
     // Generate d.ts files
     "declaration": true,
     // This compiler run should
