+++ vite.config.js
@@ -1,4 +1,4 @@
-// vite.config.ts
+// vite.config.js
 
 /// <reference types="vitest" />
 // Configure Vitest (https://vitest.dev/config/)
@@ -12,12 +12,26 @@ export default defineConfig({
   build: {
     lib: {
       // eslint-disable-next-line no-undef
-      entry: resolve(__dirname, 'src/index.js'),
+      entry: resolve(__dirname, 'src/composite.js'),
       name: 'http-please',
-      formats: ['es', 'umd'],
-      fileName: 'http-please',
+      formats: ['es'],
+      fileName: 'index',
     },
   },
-  plugins: [dts()],
+  plugins: [
+    dts({
+      include: ['src/composite.js'],
+      beforeWriteFile: (filePath, content) => {
+        return {
+          filePath: filePath.replace(
+            // eslint-disable-next-line no-undef
+            'dist/composite.d.ts',
+            'dist/index.d.ts'
+          ),
+          content,
+        };
+      },
+    }),
+  ],
   test: {},
 });
